Add ability to restart game
It is now possible to restart the game, however the functionality is
incredibly glitchy due to a faulty design choice with the frame class.
I have fixed this design flaw beginning with the Galaga+ game I'm working
on currently and understand how to avoid the problem while working on
projects in the future.